Finalizes check answer features

+++ db/migrate/20140205183644_create_take_quizzes.rb
@@ -1,7 +1,7 @@
 class CreateTakeQuizzes < ActiveRecord::Migration
   def change
     create_table :take_quizzes do |t|
-      t.integer :answer
+      t.decimal :answer
       t.belongs_to :school
 
       t.timestamps
